✨ 댓글 추가 로직 분리
