+++ src/feature/addComment/ui/AddComment.tsx
@@ -1,31 +1,26 @@
+import { addComment } from '@/entities/comment';
 import {
   useCommentActions,
   useComments,
   useDialogActions,
   useNewComment,
-  useShowAddCommentDialog,
 } from '@/shared/model/store';
-import { Button, Dialog, DialogContent, DialogHeader, DialogTitle, Textarea } from '@/shared/ui';
+import { Button } from '@/shared/ui';
 
 export const AddComment = () => {
   const comments = useComments();
   const newComment = useNewComment();
-  const { setComments, setNewComment } = useCommentActions();
 
-  const showAddCommentDialog = useShowAddCommentDialog();
+  const { setComments, setNewComment } = useCommentActions();
   const { setShowAddCommentDialog } = useDialogActions();
 
-  const addComment = async () => {
+  const handleClickAdd = async () => {
     try {
-      const response = await fetch('/api/comments/add', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(newComment),
-      });
-      const data = await response.json();
+      const { data: commentRes } = await addComment(newComment);
+
       setComments({
         ...comments,
-        [data.postId]: [...(comments[data.postId] || []), data],
+        [commentRes.postId]: [...(comments[commentRes.postId] || []), commentRes],
       });
 
       setShowAddCommentDialog(false);
@@ -35,24 +30,5 @@ export const AddComment = () => {
     }
   };
 
-  return (
-    <Dialog
-      open={showAddCommentDialog}
-      onOpenChange={setShowAddCommentDialog}
-    >
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>새 댓글 추가</DialogTitle>
-        </DialogHeader>
-        <div className='space-y-4'>
-          <Textarea
-            placeholder='댓글 내용'
-            value={newComment.body}
-            onChange={(e) => setNewComment({ ...newComment, body: e.target.value })}
-          />
-          <Button onClick={addComment}>댓글 추가</Button>
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
+  return <Button onClick={handleClickAdd}>댓글 추가</Button>;
 };
